Fix return value of poll_host_device
When receiving frames and no further frames are available, read() returns
EAGAIN. In the previous implementation, frames would not be processed until
the circular multicast buffer is full, resulting in a periodic burst of
packets. In general, a receive failure should not affect processing of packets
that were already successfully received.

Tested on:
- TP-Link Archer T1U V1.0 (mt76)
- Alfa AWUS036NHA (ath9k_htc)